backend/requirements-dev.txt
fastapi-cli[standard]
uvicorn[standard]

supabase
python-dotenv
